qt602240_ts: clear unused slots in the framework after calibration
It is possible for the state of the framework and the driver to get out
of sync resulting in "stuck" touch points from the framework's perspective.
This seems to happen only if there are touches during the device's
calibration.  The result is a locked up touchscreen for the user.

After error cases and after calibration, send an "unused" message
(-1 for ABS_MT_TRACKING_ID) for all unused slots. This fixes the
phantom stuck points.

Change-Id: I1fc22175dde9deb68ac69c322e305fc174fdba78
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>