msm8x60_celox: Add panel undervolt interface
Panel voltage can be changed on-the-go via the sysfs file /sys/module/board_msm8x60_celox/parameters/panel_uv, where panel_uv is a value between 0 and 500mV to undervolt the panel by (stock panel voltage is 3000mV). After the sysfs control is written to with the desired undervolt value (as a multiple of 25mV), turn the screen off and on to apply the new voltage.

Default panel undervolt is 300mV. This interface is currently only configured for Skyrocket.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>