mdp4_overlay: Fix graphical glitches caused by iommu mapping
DISPLAY_READ_DOMAIN 2 & GEN_POOL 2 aren't being unmapped correctly, so when mdp tries to map it, it gives an error that it is already mapped and causes visual glitches. Therefore, always unmap iommu for DISPLAY_READ_DOMAIN 2 & GEN_POOL 2 before mapping iommu for it. Run this in a loop to ensure sure that ion_map_iommu() is always successful.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>