msm: cpufreq: Sync maxfreq of all CPUs with CPU0
When a CPU core is plugged back in, the maxfreq of the plugged core resets to the maxfreq it had set at init time, thus losing any custom maxfreq set for that core. Sync the maxfreq of all cores with core0 to fix this issue.

Thanks to @EmmanuelU for reporting this issue.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>